Add .reader.xml.common for shared code
